pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sepa_ai"
version = "1.0.0"
description = "Minervini SEPA Stock Analysis System for NSE / Indian Equities"
readme = "PROJECT_DESIGN.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
dependencies = [
    # --- data acquisition -------------------------------------------------------
    "yfinance>=0.2.40",
    "nsepython>=2.10",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    # --- data processing --------------------------------------------------------
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "pyarrow>=15.0.0",
    "pandas-market-calendars>=4.4.0",
    "openpyxl>=3.1.0",
    # --- config / validation ----------------------------------------------------
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    # --- API --------------------------------------------------------------------
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "slowapi>=0.1.9",
    # --- dashboard --------------------------------------------------------------
    "streamlit>=1.33.0",
    # --- charting ---------------------------------------------------------------
    "mplfinance>=0.12.10b0",
    # --- scheduling -------------------------------------------------------------
    "apscheduler>=3.10.0",
    # --- alerts -----------------------------------------------------------------
    "python-telegram-bot>=21.0",
    # --- templating -------------------------------------------------------------
    "jinja2>=3.1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "black>=24.3.0",
    "mypy>=1.9.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = [
    "ingestion*", "features*", "rules*", "screener*",
    "paper_trading*", "llm*", "pipeline*", "backtest*",
    "alerts*", "reports*", "storage*", "utils*",
    "api*", "dashboard*",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (bootstrap/full-pipeline tests; deselect with '-m not slow')",
    "benchmark: performance benchmarks with wall-clock thresholds",
]
filterwarnings = [
    # Starlette internally references the old HTTP_422_UNPROCESSABLE_ENTITY constant
    # name (renamed to HTTP_422_UNPROCESSABLE_CONTENT in newer Starlette).  This is
    # a library-internal deprecation — not triggered by our code — and will resolve
    # when the Starlette version bundled with FastAPI is updated.
    "ignore:'HTTP_422_UNPROCESSABLE_ENTITY' is deprecated:DeprecationWarning:starlette",
]
